10403-moved services specs file under tests folder-AB#10403 AB#11013
