+++ Tests/UnitTests/fss-search-helper.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { FssSearchHelperService } from './fss-search-helper.service';
+import { FssSearchHelperService } from '../../src/app/core/services/fss-search-helper.service';
 
 describe('FssSearchHelperService', () => {
   let service: FssSearchHelperService;
